requirements.txt
langchain
qdrant_client
tiktoken
huggingface_hub
accelerate
openai
python-dotenv
pypdf
sentence-transformers
streamlit